# openapi.yaml
openapi: 3.1.0
info:
  title: DataShare API
  description: |
    API for freelance and small business file sharing.
    It allows users to register, upload (private or public), retrieve an
    upload history, generate short lived download tokens and download files
    either directly or after a password protected unlock flow.
  version: 1.0.0
servers:
  - url: https://datashare.io
    description: Production server
  - url: http://localhost:8080
    description: Local development server (Spring Boot)
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: alice@example.com
              password: secret123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request (e.g., duplicate email)
  /auth/login:
    post:
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: alice@example.com
              password: secret123
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
        userId:
          type: string
          format: uuid
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        expiresIn:
          type: string
          description: Duration before token expires (e.g. "3600s")

    # -----------------------------------------------------------------
    #  Upload file DTO – used for both private and public uploads
    # -----------------------------------------------------------------
...